# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CRM Comercial B2B API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /salespeople:
    get:
      operationId: getSalespeople
      tags: [salespeople]
      summary: List salespeople
      responses:
        "200":
          description: List of salespeople
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Salesperson"
    post:
      operationId: createSalesperson
      tags: [salespeople]
      summary: Create salesperson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalespersonRequest"
      responses:
        "201":
          description: Salesperson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salesperson"
  /salespeople/{id}:
    get:
      operationId: getSalesperson
      tags: [salespeople]
      summary: Get salesperson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Salesperson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salesperson"
    patch:
      operationId: updateSalesperson
      tags: [salespeople]
      summary: Update salesperson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSalespersonRequest"
      responses:
        "200":
          description: Updated salesperson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salesperson"
    delete:
      operationId: deleteSalesperson
      tags: [salespeople]
      summary: Delete salesperson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /opportunities:
    get:
      operationId: getOpportunities
      tags: [opportunities]
      summary: List opportunities
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: clientId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of opportunities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityListResponse"
    post:
      operationId: createOpportunity
      tags: [opportunities]
      summary: Create opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpportunityRequest"
      responses:
        "201":
          description: Opportunity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
  /opportunities/{id}:
    get:
      operationId: getOpportunity
      tags: [opportunities]
      summary: Get opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
    patch:
      operationId: updateOpportunity
      tags: [opportunities]
      summary: Update opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOpportunityRequest"
      responses:
        "200":
          description: Updated opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
...
  /prompts:
    get:
      operationId: getPrompts
      tags: [prompts]
      summary: List AI prompts
      responses:
        "200":
          description: List of prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prompt"
    post:
      operationId: createPrompt
      tags: [prompts]
      summary: Create prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromptRequest"
      responses:
        "201":
          description: Prompt created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /prompts/{id}:
    patch:
      operationId: updatePrompt
      tags: [prompts]
      summary: Update prompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromptRequest"
      responses:
        "200":
          description: Updated prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /prompts/{id}/activate:
    post:
      operationId: activatePrompt
      tags: [prompts]
      summary: Activate a prompt version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prompt activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /prompts/{id}/test:
    post:
      operationId: testPrompt
      tags: [prompts]
      summary: Test a prompt with sample text
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestPromptRequest"
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestPromptResult"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Salesperson:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        territory:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - isActive
        - createdAt
    CreateSalespersonRequest:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        territory:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - email
    UpdateSalespersonRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        territory:
          type: string
        isActive:
          type: boolean
    Product:
      type: object
      properties:
        id:
          type: integer
        code:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        dimensions:
          type: ["string", "null"]
        standard:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - isActive
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        unit:
          type: string
        category:
          type: string
        dimensions:
          type: string
        standard:
          type: string
        price:
          type: number
        currency:
          type: string
        isActive:
          type: boolean
      required:
        - name
...
    Opportunity:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        clientId:
          type: ["integer", "null"]
        contactId:
          type: ["integer", "null"]
        salespersonId:
          type: ["integer", "null"]
        emailId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [new, quoted, negotiating, won, lost, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        estimatedValue:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        extractedProducts:
          type: ["object", "null"]
        nextFollowUpAt:
          type: ["string", "null"]
          format: date-time
        closedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - priority
        - createdAt
        - updatedAt
    OpportunityListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Opportunity"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateOpportunityRequest:
      type: object
      properties:
        title:
          type: string
        clientId:
          type: integer
        contactId:
          type: integer
        salespersonId:
          type: integer
        emailId:
          type: integer
        status:
          type: string
          enum: [new, quoted, negotiating, won, lost, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        estimatedValue:
          type: number
        currency:
          type: string
        description:
          type: string
        nextFollowUpAt:
          type: string
          format: date-time
      required:
        - title
    UpdateOpportunityRequest:
      type: object
      properties:
        title:
          type: string
        clientId:
          type: integer
        contactId:
          type: integer
        salespersonId:
          type: integer
        status:
          type: string
          enum: [new, quoted, negotiating, won, lost, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        estimatedValue:
          type: number
        currency:
          type: string
        description:
          type: string
        nextFollowUpAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
...
    Prompt:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [classification, extraction, reply_draft]
        content:
          type: string
        version:
          type: integer
        isActive:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - content
        - version
        - isActive
        - createdAt
    CreatePromptRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [classification, extraction, reply_draft]
        content:
          type: string
        notes:
          type: string
      required:
        - name
        - type
        - content
    TestPromptRequest:
      type: object
      properties:
        sampleText:
          type: string
      required:
        - sampleText
    TestPromptResult:
      type: object
      properties:
        result:
          type: object
        rawOutput:
          type: string
      required:
        - result
        - rawOutput